requirements.txt
feedparser==5.2.1
pymongo==3.3.0
Unidecode==0.4.20
tweepy==3.5.0
python-decouple==3.0
TextBlob==0.12.0